requirements.txt
fastapi>=0.111
uvicorn[standard]
pandas>=2.2
scikit-learn==1.6.1
catboost==1.2.8
joblib
pydantic-settings
python-dotenv
